pet finished, order/store started

+++ src/main/java/com/model/Category.java
@@ -1,22 +1,22 @@
 package com.model;
 
 public class Category {
-    private Integer id;
+    private Long id;
     private String name;
 
     public Category() {
     }
 
-    public Category(Integer id, String name) {
+    public Category(Long id, String name) {
         this.id = id;
         this.name = name;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
