+++ src/main/java/com/model/Pet.java
@@ -3,7 +3,7 @@ package com.model;
 import java.util.List;
 
 public class Pet {
-    private Integer id;
+    private Long id;
     private Category category;
     private String name;
     private List<String> photoUrls;
@@ -29,7 +29,7 @@ public class Pet {
     public Pet() {
     }
 
-    public Pet(Integer id, Category category, String name, List<String> photoUrls, List<Tag> tags, String status) {
+    public Pet(Long id, Category category, String name, List<String> photoUrls, List<Tag> tags, String status) {
         this.id = id;
         this.category = category;
         this.name = name;
@@ -38,11 +38,11 @@ public class Pet {
         this.status = status;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
